Made the code modular
